settle behave step def discovery and project structure
