+++ tests/acc/features/steps/unboilerplate_steps.py
@@ -1,0 +1,23 @@
+'''
+Created on Jan 23, 2018
+
+@author: thomasroper
+
+	Copyright 2018 Unspecified LLC
+	Licensed under the Apache License, Version 2.0 (the "License");
+	you may not use this file except in compliance with the License.
+	You may obtain a copy of the License at
+	
+	http://www.apache.org/licenses/LICENSE-2.0
+
+	Unless required by applicable law or agreed to in writing, software
+	distributed under the License is distributed on an "AS IS" BASIS,
+	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+	See the License for the specific language governing permissions and
+	limitations under the License.
+'''
+from behave import * 
+	
+@when(u'I rock the mic like a vandal"')
+def step_impl(context, value):
+	pass
